ByteTrack-YOLO/requirements.txt
# ByteTrack-YOLO Requirements

# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
scipy>=1.7.0
lap>=0.4.0

# Deep learning
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Configuration
PyYAML>=6.0

# GUI (optional)
pillow>=9.0.0

# Development (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
